requirements.txt
streamlit>=1.39.0
numpy>=1.26.0
pandas>=2.2.0
opencv-python>=4.9.0
pillow>=10.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-image>=0.24.0
scikit-learn>=1.6.0
joblib